better logic for menu, fixes in display and after edit modal
