+++ components/Menu/Menu.js
@@ -2,6 +2,7 @@ import What from "../What/What.js";
 import { whatContent } from "../../textContent.js";
 import ListContainer from "../ListContainer/ListContainer.js";
 import Analysis from "../Analysis/Analysis.js";
+import { deActivate, removeAndHideEverything } from "../../menuLogic.js";
 
 export default function Menu() {
   const menu = document.createElement("nav");
@@ -10,7 +11,7 @@ export default function Menu() {
     <button id="what-button" class="menu_button" data-js="what-button">What?</button>
     <button id="new-task-button" class="menu_button--active" data-js="new-task-button">New</button>
     <button id="done-tasks-button" class="menu_button" data-js="done-tasks-button">Done</button>
-     <button id="analysis-button" class="menu_button" data-js="analysis-button">Analysis</button>
+    <button id="analysis-button" class="menu_button" data-js="analysis-button">Analysis</button>
 `;
   const whatButton = menu.querySelector('[data-js="what-button"]');
   const newTaskButton = menu.querySelector('[data-js="new-task-button"]');
@@ -24,44 +25,19 @@ export default function Menu() {
 
   const userEntries = JSON.parse(localStorage.getItem("RecordedTasks")) || [];
 
-  function removeAndHideEverything() {
-    const what = document.getElementById("what");
-    const newTask = document.getElementById("form-container");
-    const edit = document.getElementById("edit-container");
-    const listContainer = document.getElementById("list-container");
-    const analysis = document.getElementById("analysis");
-    analysis && analysis.remove();
-    listContainer && listContainer.remove();
-    what && what.remove();
-    edit && edit.remove();
-    newTask.classList.add("form_container--noDisplay");
-  }
-
-  function deActivate() {
-    const testElements = menu.getElementsByClassName("menu_button--active");
-    const testButtons = Array.prototype.filter.call(
-      testElements,
-      (testElement) => testElement.nodeName === "BUTTON"
-    );
-    const testButton = testButtons[0];
-    testButton.classList.toggle("menu_button--active");
-  }
-
   function handleClick(element) {
     element = element || window.event;
     const target = element.target || element.srcElement,
       text = target.id;
     const app = document.getElementById("app");
-    const what = document.getElementById("what");
     const newTask = document.getElementById("form-container");
     const listContainer = document.getElementById("list-container");
-    const analysis = document.getElementById("analysis");
     removeAndHideEverything();
     deActivate();
     switch (text) {
       case "what-button":
-        !what && whatButton.classList.toggle("menu_button--active");
-        !what && app.append(What(whatContent));
+        whatButton.classList.add("menu_button--active");
+        app.append(What(whatContent));
         break;
       case "new-task-button":
         newTaskButton.classList.add("menu_button--active");
@@ -79,11 +55,9 @@ export default function Menu() {
         }
         break;
       case "analysis-button":
-        !analysis && analysisButton.classList.toggle("menu_button--active");
+        analysisButton.classList.add("menu_button--active");
         const myUserEntries =
           JSON.parse(localStorage.getItem("RecordedTasks")) || [];
-        if (!analysis && myUserEntries.length != userEntries.length) {
-        }
         app.append(Analysis(myUserEntries));
         break;
     }
